// src/app/types/product.type.ts
export interface Product {
    intent: string;
    name: string;
    type: string;
    brand: string;
    quantity: string;
    condition: string;
    price: string;
    remarks: string | null;
    ram: string;
    color: string;
    storage: string;
    processor: string;
    chatName: string;
    message: string;
    author: string;
    messagetime: string;
}